First pass at required changes
