+++ CMakeLists.txt
@@ -31,8 +31,6 @@ pico_add_extra_outputs(${PROJECT_NAME})
 
 target_link_libraries(${PROJECT_NAME}
     pico_stdlib
-    hardware_pio
-    pico_multicore
     sparselife
     leddriver
 )
