+++ leddriver/src/utility.cpp
@@ -1,0 +1,23 @@
+#include "utility.hpp"
+
+void core1Entry()
+{
+    uint32_t fifo_value = multicore_fifo_pop_blocking();
+    LEDDriver::LEDArray *ledArr = reinterpret_cast<LEDDriver::LEDArray *>(fifo_value);
+
+    const uint32_t refreshms = multicore_fifo_pop_blocking();
+
+    while (true)
+    {
+        auto targetTime = make_timeout_time_ms(refreshms);
+        ledArr->SendBuffer();
+        sleep_until(targetTime);
+    }
+}
+
+void LaunchOnCore1(LEDDriver::LEDArray *ledArr, const uint32_t refreshms)
+{
+    multicore_launch_core1(core1Entry);
+    multicore_fifo_push_blocking(reinterpret_cast<uint32_t>(ledArr));
+    multicore_fifo_push_blocking(refreshms);
+}
